refactor to merge similiar codes as a new function issues
