+++ routers/repo/milestone.go
@@ -5,23 +5,244 @@
 package repo
 
 import (
-	"bytes"
-	"strings"
+	"time"
 
 	"code.gitea.io/gitea/models"
+	"code.gitea.io/gitea/modules/auth"
 	"code.gitea.io/gitea/modules/base"
 	"code.gitea.io/gitea/modules/context"
-	"code.gitea.io/gitea/modules/indexer"
+	"code.gitea.io/gitea/modules/markup/markdown"
 	"code.gitea.io/gitea/modules/setting"
 	"code.gitea.io/gitea/modules/util"
-
-	"github.com/Unknwon/com"
 	"github.com/Unknwon/paginater"
 )
 
 const (
+	tplMilestone       base.TplName = "repo/issue/milestones"
+	tplMilestoneNew    base.TplName = "repo/issue/milestone_new"
+	tplMilestoneEdit   base.TplName = "repo/issue/milestone_edit"
 	tplMilestoneIssues base.TplName = "repo/issue/milestone_issues"
 )
+
+// Milestones render milestones page
+func Milestones(ctx *context.Context) {
+	ctx.Data["Title"] = ctx.Tr("repo.milestones")
+	ctx.Data["PageIsIssueList"] = true
+	ctx.Data["PageIsMilestones"] = true
+
+	isShowClosed := ctx.Query("state") == "closed"
+	openCount, closedCount, err := models.MilestoneStats(ctx.Repo.Repository.ID)
+	if err != nil {
+		ctx.ServerError("MilestoneStats", err)
+		return
+	}
+	ctx.Data["OpenCount"] = openCount
+	ctx.Data["ClosedCount"] = closedCount
+
+	sortType := ctx.Query("sort")
+	page := ctx.QueryInt("page")
+	if page <= 1 {
+		page = 1
+	}
+
+	var total int
+	if !isShowClosed {
+		total = int(openCount)
+	} else {
+		total = int(closedCount)
+	}
+	ctx.Data["Page"] = paginater.New(total, setting.UI.IssuePagingNum, page, 5)
+
+	miles, err := models.GetMilestones(ctx.Repo.Repository.ID, page, isShowClosed, sortType)
+	if err != nil {
+		ctx.ServerError("GetMilestones", err)
+		return
+	}
+	if ctx.Repo.Repository.IsTimetrackerEnabled() {
+		if miles.LoadTotalTrackedTimes(); err != nil {
+			ctx.ServerError("LoadTotalTrackedTimes", err)
+			return
+		}
+	}
+	for _, m := range miles {
+		m.RenderedContent = string(markdown.Render([]byte(m.Content), ctx.Repo.RepoLink, ctx.Repo.Repository.ComposeMetas()))
+	}
+	ctx.Data["Milestones"] = miles
+
+	if isShowClosed {
+		ctx.Data["State"] = "closed"
+	} else {
+		ctx.Data["State"] = "open"
+	}
+
+	ctx.Data["SortType"] = sortType
+	ctx.Data["IsShowClosed"] = isShowClosed
+	ctx.HTML(200, tplMilestone)
+}
+
+// NewMilestone render creating milestone page
+func NewMilestone(ctx *context.Context) {
+	ctx.Data["Title"] = ctx.Tr("repo.milestones.new")
+	ctx.Data["PageIsIssueList"] = true
+	ctx.Data["PageIsMilestones"] = true
+	ctx.Data["RequireDatetimepicker"] = true
+	ctx.Data["DateLang"] = setting.DateLang(ctx.Locale.Language())
+	ctx.HTML(200, tplMilestoneNew)
+}
+
+// NewMilestonePost response for creating milestone
+func NewMilestonePost(ctx *context.Context, form auth.CreateMilestoneForm) {
+	ctx.Data["Title"] = ctx.Tr("repo.milestones.new")
+	ctx.Data["PageIsIssueList"] = true
+	ctx.Data["PageIsMilestones"] = true
+	ctx.Data["RequireDatetimepicker"] = true
+	ctx.Data["DateLang"] = setting.DateLang(ctx.Locale.Language())
+
+	if ctx.HasError() {
+		ctx.HTML(200, tplMilestoneNew)
+		return
+	}
+
+	if len(form.Deadline) == 0 {
+		form.Deadline = "9999-12-31"
+	}
+	deadline, err := time.ParseInLocation("2006-01-02", form.Deadline, time.Local)
+	if err != nil {
+		ctx.Data["Err_Deadline"] = true
+		ctx.RenderWithErr(ctx.Tr("repo.milestones.invalid_due_date_format"), tplMilestoneNew, &form)
+		return
+	}
+
+	if err = models.NewMilestone(&models.Milestone{
+		RepoID:       ctx.Repo.Repository.ID,
+		Name:         form.Title,
+		Content:      form.Content,
+		DeadlineUnix: util.TimeStamp(deadline.Unix()),
+	}); err != nil {
+		ctx.ServerError("NewMilestone", err)
+		return
+	}
+
+	ctx.Flash.Success(ctx.Tr("repo.milestones.create_success", form.Title))
+	ctx.Redirect(ctx.Repo.RepoLink + "/milestones")
+}
+
+// EditMilestone render edting milestone page
+func EditMilestone(ctx *context.Context) {
+	ctx.Data["Title"] = ctx.Tr("repo.milestones.edit")
+	ctx.Data["PageIsMilestones"] = true
+	ctx.Data["PageIsEditMilestone"] = true
+	ctx.Data["RequireDatetimepicker"] = true
+	ctx.Data["DateLang"] = setting.DateLang(ctx.Locale.Language())
+
+	m, err := models.GetMilestoneByRepoID(ctx.Repo.Repository.ID, ctx.ParamsInt64(":id"))
+	if err != nil {
+		if models.IsErrMilestoneNotExist(err) {
+			ctx.NotFound("", nil)
+		} else {
+			ctx.ServerError("GetMilestoneByRepoID", err)
+		}
+		return
+	}
+	ctx.Data["title"] = m.Name
+	ctx.Data["content"] = m.Content
+	if len(m.DeadlineString) > 0 {
+		ctx.Data["deadline"] = m.DeadlineString
+	}
+	ctx.HTML(200, tplMilestoneNew)
+}
+
+// EditMilestonePost response for edting milestone
+func EditMilestonePost(ctx *context.Context, form auth.CreateMilestoneForm) {
+	ctx.Data["Title"] = ctx.Tr("repo.milestones.edit")
+	ctx.Data["PageIsMilestones"] = true
+	ctx.Data["PageIsEditMilestone"] = true
+	ctx.Data["RequireDatetimepicker"] = true
+	ctx.Data["DateLang"] = setting.DateLang(ctx.Locale.Language())
+
+	if ctx.HasError() {
+		ctx.HTML(200, tplMilestoneNew)
+		return
+	}
+
+	if len(form.Deadline) == 0 {
+		form.Deadline = "9999-12-31"
+	}
+	deadline, err := time.ParseInLocation("2006-01-02", form.Deadline, time.Local)
+	if err != nil {
+		ctx.Data["Err_Deadline"] = true
+		ctx.RenderWithErr(ctx.Tr("repo.milestones.invalid_due_date_format"), tplMilestoneNew, &form)
+		return
+	}
+
+	m, err := models.GetMilestoneByRepoID(ctx.Repo.Repository.ID, ctx.ParamsInt64(":id"))
+	if err != nil {
+		if models.IsErrMilestoneNotExist(err) {
+			ctx.NotFound("", nil)
+		} else {
+			ctx.ServerError("GetMilestoneByRepoID", err)
+		}
+		return
+	}
+	m.Name = form.Title
+	m.Content = form.Content
+	m.DeadlineUnix = util.TimeStamp(deadline.Unix())
+	if err = models.UpdateMilestone(m); err != nil {
+		ctx.ServerError("UpdateMilestone", err)
+		return
+	}
+
+	ctx.Flash.Success(ctx.Tr("repo.milestones.edit_success", m.Name))
+	ctx.Redirect(ctx.Repo.RepoLink + "/milestones")
+}
+
+// ChangeMilestonStatus response for change a milestone's status
+func ChangeMilestonStatus(ctx *context.Context) {
+	m, err := models.GetMilestoneByRepoID(ctx.Repo.Repository.ID, ctx.ParamsInt64(":id"))
+	if err != nil {
+		if models.IsErrMilestoneNotExist(err) {
+			ctx.NotFound("", err)
+		} else {
+			ctx.ServerError("GetMilestoneByRepoID", err)
+		}
+		return
+	}
+
+	switch ctx.Params(":action") {
+	case "open":
+		if m.IsClosed {
+			if err = models.ChangeMilestoneStatus(m, false); err != nil {
+				ctx.ServerError("ChangeMilestoneStatus", err)
+				return
+			}
+		}
+		ctx.Redirect(ctx.Repo.RepoLink + "/milestones?state=open")
+	case "close":
+		if !m.IsClosed {
+			m.ClosedDateUnix = util.TimeStampNow()
+			if err = models.ChangeMilestoneStatus(m, true); err != nil {
+				ctx.ServerError("ChangeMilestoneStatus", err)
+				return
+			}
+		}
+		ctx.Redirect(ctx.Repo.RepoLink + "/milestones?state=closed")
+	default:
+		ctx.Redirect(ctx.Repo.RepoLink + "/milestones")
+	}
+}
+
+// DeleteMilestone delete a milestone
+func DeleteMilestone(ctx *context.Context) {
+	if err := models.DeleteMilestoneByRepoID(ctx.Repo.Repository.ID, ctx.QueryInt64("id")); err != nil {
+		ctx.Flash.Error("DeleteMilestoneByRepoID: " + err.Error())
+	} else {
+		ctx.Flash.Success(ctx.Tr("repo.milestones.deletion_success"))
+	}
+
+	ctx.JSON(200, map[string]interface{}{
+		"redirect": ctx.Repo.RepoLink + "/milestones",
+	})
+}
 
 // MilestoneIssuesAndPulls lists all the issues and pull requests of the milestone
 func MilestoneIssuesAndPulls(ctx *context.Context) {
@@ -35,138 +256,7 @@ func MilestoneIssuesAndPulls(ctx *context.Context) {
 	ctx.Data["Title"] = milestone.Name
 	ctx.Data["Milestone"] = milestone
 
-	viewType := ctx.Query("type")
-	sortType := ctx.Query("sort")
-	types := []string{"all", "your_repositories", "assigned", "created_by", "mentioned"}
-	if !com.IsSliceContainsStr(types, viewType) {
-		viewType = "all"
-	}
-
-	var (
-		assigneeID  = ctx.QueryInt64("assignee")
-		posterID    int64
-		mentionedID int64
-		forceEmpty  bool
-	)
-
-	if ctx.IsSigned {
-		switch viewType {
-		case "created_by":
-			posterID = ctx.User.ID
-		case "mentioned":
-			mentionedID = ctx.User.ID
-		}
-	}
-
-	repo := ctx.Repo.Repository
-	selectLabels := ctx.Query("labels")
-
-	isShowClosed := ctx.Query("state") == "closed"
-
-	keyword := strings.Trim(ctx.Query("q"), " ")
-	if bytes.Contains([]byte(keyword), []byte{0x00}) {
-		keyword = ""
-	}
-
-	var issueIDs []int64
-	if len(keyword) > 0 {
-		issueIDs, err = indexer.SearchIssuesByKeyword(repo.ID, keyword)
-		if len(issueIDs) == 0 {
-			forceEmpty = true
-		}
-	}
-
-	var issueStats *models.IssueStats
-	if forceEmpty {
-		issueStats = &models.IssueStats{}
-	} else {
-		issueStats, err = models.GetIssueStats(&models.IssueStatsOptions{
-			RepoID:      repo.ID,
-			Labels:      selectLabels,
-			MilestoneID: milestoneID,
-			AssigneeID:  assigneeID,
-			MentionedID: mentionedID,
-			PosterID:    posterID,
-			IssueIDs:    issueIDs,
-		})
-		if err != nil {
-			ctx.ServerError("GetIssueStats", err)
-			return
-		}
-	}
-	page := ctx.QueryInt("page")
-	if page <= 1 {
-		page = 1
-	}
-
-	var total int
-	if !isShowClosed {
-		total = int(issueStats.OpenCount)
-	} else {
-		total = int(issueStats.ClosedCount)
-	}
-	pager := paginater.New(total, setting.UI.IssuePagingNum, page, 5)
-	ctx.Data["Page"] = pager
-
-	var issues []*models.Issue
-	if forceEmpty {
-		issues = []*models.Issue{}
-	} else {
-		issues, err = models.Issues(&models.IssuesOptions{
-			RepoIDs:     []int64{repo.ID},
-			AssigneeID:  assigneeID,
-			PosterID:    posterID,
-			MentionedID: mentionedID,
-			MilestoneID: milestoneID,
-			Page:        pager.Current(),
-			PageSize:    setting.UI.IssuePagingNum,
-			IsClosed:    util.OptionalBoolOf(isShowClosed),
-			Labels:      selectLabels,
-			SortType:    sortType,
-			IssueIDs:    issueIDs,
-		})
-		if err != nil {
-			ctx.ServerError("Issues", err)
-			return
-		}
-	}
-
-	// Get posters.
-	for i := range issues {
-		// Check read status
-		if !ctx.IsSigned {
-			issues[i].IsRead = true
-		} else if err = issues[i].GetIsRead(ctx.User.ID); err != nil {
-			ctx.ServerError("GetIsRead", err)
-			return
-		}
-	}
-	ctx.Data["Issues"] = issues
-
-	// Get assignees.
-	ctx.Data["Assignees"], err = repo.GetAssignees()
-	if err != nil {
-		ctx.ServerError("GetAssignees", err)
-		return
-	}
-
-	if ctx.QueryInt64("assignee") == 0 {
-		assigneeID = 0 // Reset ID to prevent unexpected selection of assignee.
-	}
-
-	ctx.Data["IssueStats"] = issueStats
-	ctx.Data["SelectLabels"] = com.StrTo(selectLabels).MustInt64()
-	ctx.Data["ViewType"] = viewType
-	ctx.Data["SortType"] = sortType
-	ctx.Data["MilestoneID"] = milestoneID
-	ctx.Data["AssigneeID"] = assigneeID
-	ctx.Data["IsShowClosed"] = isShowClosed
-	ctx.Data["Keyword"] = keyword
-	if isShowClosed {
-		ctx.Data["State"] = "closed"
-	} else {
-		ctx.Data["State"] = "open"
-	}
+	issues(ctx, milestoneID, util.OptionalBoolNone)
 
 	ctx.HTML(200, tplMilestoneIssues)
 }
